requirements.txt
flask==3.0.1
pyyaml==6.0.1